[REF] account_vat_period_end_statement: Allow to re-use test methods
(cherry picked from commit 5e96c4368f38c60f337959b43a89b88d10e06f78)
